requirements.txt
Authlib==0.14.1
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Login==0.5.0
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
marshmallow-sqlalchemy==0.22.3
nose==1.3.7
psycopg2==2.8.4
pytest==5.4.1
python-dotenv==0.12.0
tldextract==2.2.2